extract outbound delivery task
